requirements.txt
aerich==0.8.2
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8.0
attrs==25.3.0
certifi==2025.4.26
click==8.1.8
dictdiffer==0.9.0
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
idna==3.10
iso8601==2.1.0
magic-filter==1.0.12
multidict==6.4.3
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pypika-tortoise==0.5.0
python-dotenv==1.1.0
pytz==2025.2
sniffio==1.3.1
starlette==0.46.2
tortoise-orm==0.25.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
yarl==1.20.0
